move files containing manual tests into own folder
